Break out status bar icon from extdesk icon.
Create a separate resource for the dashboard category icon for
Status Bar. Rename the extended desktop icons so they are further
distinguished from the dashboard iconography.

The goal is makea sure only dashboard icons are using the
ic_settings_ prefix and rename iconography that is not part of
that view.

Ticket: CYNGNOS-1681
Change-Id: I54165bd4338bb1d73f1d453f75ee1c06cfd62a8f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: xyyx <xyyx@mail.ru>