Put code that updates the access points lists inside a Handler.
When access point is forgot, we got callbacks from WifiManager to tell
us if the forget action is successful or not. We will refresh the access
point list in both cases. However, these callbacks happens on a
different thread and updating preferences should only be done on the
main thread. Creating a handler to handler all updates in the main
thread.

Change-Id: I7593befb20e46391ad69a284375693351a2cc794
Fixes: 65745404
Test: make RunSettingsRoboTests